+++ R/sim-data-access.R
@@ -55,6 +55,11 @@
 #' @export
 see.group.data <- function(group, data.type, effect.set=1L, label=1L) {
 	if (is.null(sim.data$p)) { stop("Please load.data first.") }
+  if (!is.integer(group)) {
+    ngroup <- group
+    group <- as.integer(group)
+    if (!all.equal(ngroup,group)) { stop("Group identifiers must be integers.") }
+  }
 	return(.Call(SXP_see_group_data, sim.data$p, group, toupper(data.type), effect.set, label))
 }
 
@@ -76,7 +81,11 @@ see.group.data <- function(group, data.type, effect.set=1L, label=1L) {
 #' @export
 see.group.gene.data <- function(group, count.allele=NA_character_) {
   if (is.null(sim.data$p)) { stop("Please load.data first.") }
-  
+  if (!is.integer(group)) {
+    ngroup <- group
+    group <- as.integer(group)
+    if (!all.equal(ngroup,group)) { stop("Group identifiers must be integers.") }
+  }
   return(.Call(SXP_see_group_gene_data, sim.data$p, group, count.allele))
 }