+++ R/sim-group-utils.R
@@ -1,6 +1,6 @@
 #' Assign multiple groups' worth of genotypes to a single group
 #'
-#' \code{combine.group} combines the members of multple groups
+#' \code{combine.group} combines the members of multiple groups
 #' into a single group. This is useful for allowing certain crossing
 #' functions to act on these genotypes together.
 #'
@@ -12,6 +12,11 @@
 #' @export
 combine.groups <- function(groups) {
 	if (is.null(sim.data$p)) { stop("Please load.data first.") }
+  if (!is.integer(groups)) {
+    ngroup <- groups
+    groups <- as.integer(groups)
+    if (!all.equal(ngroup,groups)) { stop("Group identifiers must be integers.") }
+  }
 	return(.Call(SXP_combine_groups, sim.data$p, groups))
 }